pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyantdisplay"
version = "1.0.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "A Python application for reading data from ANT+ devices using an ANT+ USB stick"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware :: Universal Serial Bus (USB)",
    "Topic :: Scientific/Engineering",
]
keywords = ["ant+", "fitness", "heart-rate", "cycling", "usb", "sensors", "mqtt", "home-assistant"]
repository = "https://github.com/your-username/pyAntDisplay"
homepage = "https://github.com/your-username/pyAntDisplay"
dependencies = [
    "openant>=1.2.0",
    "pyyaml>=6.0.1",
    "colorama>=0.4.6",
    "pyusb>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
    "coverage>=6.0.0",
]

[project.urls]
"Homepage" = "https://github.com/Velotales/pyAntDisplay"
"Bug Reports" = "https://github.com/Velotales/pyAntDisplay/issues"
"Source" = "https://github.com/Velotales/pyAntDisplay"

[project.scripts]
pyantdisplay = "pyantdisplay.main:main"
pyantdisplay-monitor = "pyantdisplay.live_monitor:main"
pyantdisplay-scan = "pyantdisplay.device_scanner:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--verbose --tb=short"
minversion = "7.0"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]